Fix SpecStyle to not receive Prism events (#71)
* Calling super will still register for events

The parent class TestDiscovery also registers SpecStyle for Prism event. While handling the events it tries to access a nil variable when entering a class or module and thus completely break code analysis of normal ruby files.

* Added a test to verify expected SpecStyle patch behavior

* Added integration test to verify parsing of a simple ruby file

The improperly patched SpecStyle would fail when the Ruby LSP was
handling a 'textDocument/foldingRange' message (probably others too)
for a source while with module and class blocks as it was still
receiving events from Prism (which we do not want) for the RSpec
addon.